% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cranlog.R
\docType{package}
\name{cranlog}
\alias{cranlog}
\alias{cranlog-package}
\title{Reading and Formatting CRAN Logs}
\description{
This package provides functions to handle the
\href{http://cran-logs.rstudio.com/}{CRAN download logs},
seamlessly reading them in as a data.frame and tbl_df
through \code{readr} and simplifying various columns
if desired.
}
\seealso{
\code{\link{read_logs}}, \code{\link{format_logs}}.
}

