% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.R
\name{format_logs}
\alias{format_logs}
\title{Format Downloaded CRAN Logs}
\usage{
format_logs(logs)
}
\arguments{
\item{logs}{a data.frame of log data obtained through \code{read_logs}.}
}
\value{
a data.frame identical to \code{logs} but with a reformatted \code{r_os} field.
}
\description{
the CRAN logs are in a highly sensible format but some of the fields
haven't had their values abstracted away to easily human-readable ones. \code{format_logs}
performs this operation, starting with refactoring the \code{r_os} field to contain values
like "Windows" or "OSX" instead of "mingw32" or "darwin".
}
\examples{
# Read in and format the logs from the day before yesterday.
\dontrun{
formatted_data <- format_logs(read_logs())
}
}

