% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_logs}
\alias{read_logs}
\title{Read CRAN Download Logs}
\usage{
read_logs(start = (Sys.Date() - 2), end = NULL)
}
\arguments{
\item{start}{a Date object indicating the start of the range to read. Defaults to the day before yesterday;
CRAN logs are updated on a one-day lag and a two-day lag allows for timezone differences.}

\item{end}{a Date object indicating the end of the range to read. Defaults to NULL, in which case 'start' will
be taken as the only date to read for.}
}
\description{
foo
}
\examples{
# Read a single day's logs from 5 days ago.
data <- read_logs(start = Sys.Date() - 5)

# Read everything between 5 July and 10 July
\dontrun{
data <- read_logs(start = as.Date("2015-07-05"), end = as.Date("2015-07-10"))
}
}

